:recycle: userIdx 타입 수정

+++ src/main/java/backend/whereIsMyTeam/user/dto/UserLoginResponseDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserLoginResponseDto {
-    private Integer userIdx;
+    private Long userIdx;
     private String token;
     private String refreshToken;
 }
